+++ jest.config.js
@@ -1,7 +1,4 @@
 module.exports = {
   preset: "ts-jest",
   testEnvironment: "jsdom",
-  moduleNameMapper: {
-    "\\.(css|less)$": "<rootDir>/src/__mocks__/styleMock.js",
-  },
 };
