+++ package.json
@@ -18,6 +18,8 @@
   "devDependencies": {
     "@brave-intl/skus-sdk": "^0.1.0",
     "@types/jest": "27.0.3",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
     "autoprefixer": "10.4.0",
     "copy-webpack-plugin": "10.1.0",
     "css-loader": "6.5.1",
@@ -40,5 +42,8 @@
     "webpack-cli": "4.9.1",
     "webpack-dev-server": "4.6.0"
   },
-  "dependencies": {}
+  "dependencies": {
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
+  }
 }
