--- src/__mocks__/styleMock.js
@@ -1,2 +1,0 @@
-// see https://jestjs.io/docs/webpack#handling-static-assets
-module.exports = {};
