+++ src/css/recordings.css
@@ -14,6 +14,24 @@
     font-weight: 600;
     font-size: 22px;
     line-height: 26px;
+    margin-bottom: 33px;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 16px;
+  }
+
+  & .recordings-footer {
+    text-align: center;
+    margin-top: 35px;
+    margin-bottom: 20px;
+    font-size: 14px;
+    letter-spacing: 0.01em;
+
+    &:hover {
+      text-decoration: underline;
+    }
   }
 
   & .recording {
