+++ src/index.html
@@ -133,9 +133,7 @@
                 </div>
               </div>
             </div>
-            <div class="section recordings" id="recordings">
-              <p class="recordings-header">Your recorded calls</p>
-            </div>
+            <div class="section recordings" id="recordings"></div>
             <div class="section subscribe" id="subscribe">
               <p class="subscribe-text">
                 Upgrade to host video calls with hundreds of participants.
