+++ tsconfig.json
@@ -4,7 +4,7 @@
     "strict": true,
     "module": "ES2020",
     "target": "ES2017",
-    "jsx": "react",
+    "jsx": "react-jsx",
     "allowJs": true,
     "downlevelIteration": true,
     "moduleResolution": "node",
